将 BaseScanner 与 TokenScanner 合并
