+++ metasequoia_sql/common/__init__.py
@@ -3,4 +3,4 @@
 """
 
 from metasequoia_sql.common.basic import ordered_distinct
-from metasequoia_sql.common.token_scanner import TokenScanner
+from metasequoia_sql.common.scanner import TokenScanner
